Add in progress work
